requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.3
django-cors-headers==3.5.0
django-extensions==3.1.0
djangorestframework==3.12.2
entrypoints==0.3
gunicorn==20.0.4
idna==2.10
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
Markdown==3.3.3
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
packaging==20.4
pandocfilters==1.4.3
parso==0.7.1
pickleshare==0.7.5
prometheus-client==0.9.0
prompt-toolkit==3.0.8
psycopg2==2.8.6
pycparser==2.20
Pygments==2.7.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
requests==2.25.0
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.4.1
terminado==0.9.1
testpath==0.4.4
tornado==6.1
traitlets==5.0.5
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
